Exact Testing of Linear Model Coefficients

Performs exact tests on specified coefficients of a linear model object using permutation tests to generate p-values and confidence intervals around the estimated coefficients. This function can handle both studentized and non-studentized test statistics, and allows the user to specify various parameters for the test.

exactt(
  model,
  data,
  alpha = 0.05,
  variables = NULL,
  betaNull = NULL,
  nBlocks = 5,
  nPerms = NULL,
  studentize = TRUE,
  precisionToUse = NULL,
  permutation = NULL,
  randomizationDist = FALSE,
  optimize = FALSE,
  GX1 = TRUE,
  seed = 31740,
  ...
)

Arguments

model

A formula specifying the model..

data

A data frame or matrix containing the variables used in the model.

alpha

The significance level used for the hypothesis tests; defaults to 0.05.

variables

Optional; a character vector of predictor names to test. If NULL, all predictors in the model are tested.

betaNull

Optional; a numeric vector of null hypothesis values for the coefficients. Must be the same length as variables if not NULL.

nBlocks

The number of blocks to use for permutations.

nPerms

Optional; the number of permutations to perform. If NULL or greater than the number of possible permutations, all permutations are used.

studentize

Logical indicating whether to use studentized residuals for the test.

precisionToUse

The precision level for rounding off the beta values in the resulting sequence.

permutation

Optional; a specific permutation vector to rearrange order of data.

randomizationDist

Logical indicating whether to return randomization distribution for each null hypothesis value.

optimize

Logical indicating whether to optimize the ordering of the data.

GX1

Logical indicating whether to use GX1 or X1 when constructing eps_hat. Using X1 has slightly more power at slightly more computational expense.

seed

Seed used when optimizing using GA::ga(). Default is 31740.

...

Additional arguments passed to GA::ga() for optimizing power. This can include parameters like popSize, maxiter, parallel, etc., that are used to configure the genetic algorithm. Note that when sample size is large optimizing is computationally expensive and has little effect.

Value

An object of class 'et', which includes:

summary: A matrix summarizing the test results for each variable.

detailed: A list containing detailed test results for each variable.

gaResults: Optional; a list of results from the GA::ga() function, included only when power optimization is performed via genetic algorithm parameters. Each element of the list corresponds to results for one of the tested variables, containing details like the best permutations found, fitness scores, and other GA diagnostics.

call: The matched call.

Details

The function divides the data into blocks specified by nBlocks and performs permutations within these blocks to generate the null distribution of the test statistic. The user can specify a set number of permutations with nPerms, or allow the function to calculate all possible permutations if nPerms is unspecified or too large.

If studentize is TRUE, studentized residuals are used to adjust the test statistics, potentially leading to more robust inference under model misspecification.

The function allows for a high degree of customization through its parameters and can handle large datasets and complex model structures efficiently.