Calculate p-values for two sided case.

Calculate p-values for two sided case.

exactt.pval.new.reg(
  Y.temp,
  X1.temp,
  GX2.temp,
  permIndices,
  GX.indices,
  Q.X1.temp,
  QGX2.temp,
  QGX1GX2.temp,
  side,
  denominator
)

Arguments

Y.temp

The response vector for which the test is being performed.

X1.temp

A numeric column vector of the primary variable.

GX2.temp

A matrix of permuted versions of secondary variables.

permIndices

A matrix of permutation indices used in the test.

GX.indices

A matrix of permutation indices to create GX matrices.

Q.X1.temp

A numeric column vector of the primary variable annihilated by GX2.

QGX2.temp

The projection matrix that annihilates GX2.

QGX1GX2.temp

The projection matrix that annihilates the mean, GX1, and GX2.

side

A character to indicate the side of the test.

denominator

Character argument indicating how to calculate epsilon hat.