Summarize 'exactt' Objects

Summarize 'exactt' Objects

# S3 method for class 'exactt'
summary(object, ...)

Arguments

object

An object of class 'exactt' from the output from exactt function.

...

Additional arguments passed to the summary function.

alpha: A value between 0 and 1 indicating the level of significance.

Value

An object of class 'exactt', which includes:

summary: A matrix summarizing the test results for each variable.

detailed: A list containing detailed test results for each variable.

gaResults: Optional; a list of results from the GA::ga() function, included only when power optimization is performed via genetic algorithm parameters. Each element of the list corresponds to results for one of the tested variables, containing details like the best permutations found, fitness scores, and other GA diagnostics.

call: The matched call.